Consolidating experiments helper functions.
